Move Bert to models/ folder
