+++ keras_nlp/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_nlp import applications
+from keras_nlp import models
 from keras_nlp import layers
 from keras_nlp import metrics
 from keras_nlp import tokenizers
